+++ Tests/Tiny.RestClient.Tests/EtagTests.cs
@@ -1,11 +1,5 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
 using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace Tiny.RestClient.Tests
 {
@@ -55,9 +49,9 @@ namespace Tiny.RestClient.Tests
             var data = await client.GetRequest("GetTest/complex").
                 FillResponseHeaders(out Headers headers).
                 ExecuteAsync<string[]>();
-            Assert.AreEqual(data.Length, 2);
-            Assert.AreEqual(data[0], "value1");
-            Assert.AreEqual(data[1], "value2");
+            Assert.AreEqual(2, data.Length);
+            Assert.AreEqual("value1", data[0]);
+            Assert.AreEqual("value2", data[1]);
 
             var actionUri = new Uri($"{ServerUrl}GetTest/complex");
             var etag = headers["ETag"].FirstOrDefault();
@@ -66,14 +60,14 @@ namespace Tiny.RestClient.Tests
 
             var fakeData = new List<string>() { "test1", "test2" };
 
-            var json = client.Settings.Formatters.FirstOrDefault().Serialize<List<string>>(fakeData, client.Settings.Encoding);
+            var json = await client.Settings.Formatters.FirstOrDefault().SerializeAsync<List<string>>(fakeData, client.Settings.Encoding, CancellationToken.None);
             await etagContainer.SaveDataAsync(actionUri, etagStored, new MemoryStream(Encoding.UTF8.GetBytes(json)), CancellationToken.None);
 
             data = await client.GetRequest("GetTest/complex").
                 ExecuteAsync<string[]>();
-            Assert.AreEqual(data.Length, 2);
-            Assert.AreEqual(data[0], "test1");
-            Assert.AreEqual(data[1], "test2");
+            Assert.AreEqual(2, data.Length);
+            Assert.AreEqual("test1", data[0]);
+            Assert.AreEqual("test2", data[1]);
 
             await etagContainer.SaveDataAsync(actionUri, "\"TEST\"", new MemoryStream(), CancellationToken.None);
 
@@ -83,9 +77,9 @@ namespace Tiny.RestClient.Tests
             etagStored = await etagContainer.GetExistingETagAsync(actionUri, CancellationToken.None);
             Assert.AreEqual(etagStored, etag);
 
-            Assert.AreEqual(data.Length, 2);
-            Assert.AreEqual(data[0], "value1");
-            Assert.AreEqual(data[1], "value2");
+            Assert.AreEqual(2, data.Length);
+            Assert.AreEqual("value1", data[0]);
+            Assert.AreEqual("value2", data[1]);
         }
 
         [TestMethod]
@@ -98,9 +92,9 @@ namespace Tiny.RestClient.Tests
                 WithETagContainer(etagContainer).
                 FillResponseHeaders(out Headers headers).
                 ExecuteAsync<string[]>();
-            Assert.AreEqual(data.Length, 2);
-            Assert.AreEqual(data[0], "value1");
-            Assert.AreEqual(data[1], "value2");
+            Assert.AreEqual(2, data.Length);
+            Assert.AreEqual("value1", data[0]);
+            Assert.AreEqual("value2", data[1]);
 
             var actionUri = new Uri($"{ServerUrl}GetTest/complex");
             var etag = headers["ETag"].FirstOrDefault();
@@ -109,15 +103,15 @@ namespace Tiny.RestClient.Tests
 
             var fakeData = new List<string>() { "test1", "test2" };
 
-            var json = client.Settings.Formatters.FirstOrDefault().Serialize<List<string>>(fakeData, client.Settings.Encoding);
+            var json = await client.Settings.Formatters.FirstOrDefault().SerializeAsync<List<string>>(fakeData, client.Settings.Encoding, CancellationToken.None);
             await etagContainer.SaveDataAsync(actionUri, etagStored, new MemoryStream(Encoding.UTF8.GetBytes(json)), CancellationToken.None);
 
             data = await client.GetRequest("GetTest/complex").
                 WithETagContainer(etagContainer).
                 ExecuteAsync<string[]>();
-            Assert.AreEqual(data.Length, 2);
-            Assert.AreEqual(data[0], "test1");
-            Assert.AreEqual(data[1], "test2");
+            Assert.AreEqual(2, data.Length);
+            Assert.AreEqual("test1", data[0]);
+            Assert.AreEqual("test2", data[1]);
 
             await etagContainer.SaveDataAsync(actionUri, "\"TEST\"", new MemoryStream(), CancellationToken.None);
 
@@ -128,17 +122,16 @@ namespace Tiny.RestClient.Tests
             etagStored = await etagContainer.GetExistingETagAsync(actionUri, CancellationToken.None);
             Assert.AreEqual(etagStored, etag);
 
-            Assert.AreEqual(data.Length, 2);
-            Assert.AreEqual(data[0], "value1");
-            Assert.AreEqual(data[1], "value2");
+            Assert.AreEqual(2, data.Length);
+            Assert.AreEqual("value1", data[0]);
+            Assert.AreEqual("value2", data[1]);
         }
 
         [TestMethod]
-        [ExpectedException(typeof(DirectoryNotFoundException))]
         public void ETagFileContainerDirectoryNotFound()
         {
-            new ETagFileContainer(@"C:\notfound");
-            Assert.Fail("It must not go here");
+            Assert.ThrowsExactly<DirectoryNotFoundException>(() =>
+                new ETagFileContainer(@"C:\notfound"));
         }
     }
 }