+++ Tiny.RestClient/Request/IRequest.cs
@@ -4,7 +4,8 @@ using System.IO;
 namespace Tiny.RestClient
 {
     /// <summary>
-    /// Interface IRequest.
+    /// Represents a request whose body has not been set yet. Add a body (serialized content, string, stream,
+    /// byte array or file), switch to a multipart request, or execute it directly.
     /// </summary>
     /// <seealso cref="IFormRequest" />
     /// <seealso cref="IExecutableRequest" />
@@ -16,9 +17,8 @@ namespace Tiny.RestClient
         /// <typeparam name="TContent">The type of the t content.</typeparam>
         /// <param name="content">The content.</param>
         /// <param name="serializer">Override the default serializer setted on the client.</param>
-        /// <param name="compression">Add a compression system to compress your content.</param>
         /// <returns>The current request.</returns>
-        IParameterRequest AddContent<TContent>(TContent content, IFormatter serializer = null, ICompression compression = null)
+        IParameterRequest AddContent<TContent>(TContent content, IFormatter serializer = null)
             where TContent : class;
 
         /// <summary>
@@ -44,7 +44,7 @@ namespace Tiny.RestClient
         /// <param name="contentType">The content type.</param>
         /// <returns>The current request.</returns>
         IParameterRequest AddStringContent(string content, string contentType = "text/plain");
-#if !FILEINFO_NOT_SUPPORTED
+
         /// <summary>
         /// Adds file as content.
         /// </summary>
@@ -52,13 +52,21 @@ namespace Tiny.RestClient
         /// <param name="contentType">The content type.</param>
         /// <returns>The current request.</returns>
         IParameterRequest AddFileContent(FileInfo file, string contentType);
-#endif
 
         /// <summary>
-        /// As a multipart data from request.
+        /// Switches the request to a multipart/form-data request, allowing several parts (objects, byte arrays, streams, strings, files) to be added.
         /// </summary>
-        /// <param name="contentType">content type of the request (default value  = "multipart/form-data").</param>
-        /// <returns>The current request.</returns>
+        /// <param name="contentType">Content type of the request (default value = "multipart/form-data").</param>
+        /// <returns>The current request as a multipart request.</returns>
+        /// <example>
+        /// <code>
+        /// await client.PostRequest("MultiPart/Test")
+        ///     .AsMultiPartFromDataRequest()
+        ///     .AddContent&lt;City&gt;(city1, "city1", "city1.json")
+        ///     .AddContent&lt;City&gt;(city2, "city2", "city2.json")
+        ///     .ExecuteAsync();
+        /// </code>
+        /// </example>
         IMultipartFromDataRequest AsMultiPartFromDataRequest(string contentType = "multipart/form-data");
     }
 }